backend-headless: Create a new header for headless internals
Create a new header file called headless-internal.h that includes
the relevant types and other internal declarations.
